Move RequestConfirm to react/components
