+++ indico/web/client/js/react/components/RequestConfirm.jsx
