+++ indico/web/client/js/react/components/index.js
@@ -33,3 +33,4 @@ export {default as PublicationSwitch} from './PublicationSwitch';
 export {default as MathJax} from './MathJax';
 export {default as FileSubmission} from './FileSubmission';
 export {default as ClipboardButton} from './ClipboardButton';
+export {default as RequestConfirm} from './RequestConfirm';
